refactor: split ingestion capacity dtos
